Refactor code, remove duplicate interfaces and moved common service into seperate file.
