+++ composables/movie-service.ts
@@ -1,34 +1,21 @@
 import { API_BASE_URL, API_KEY } from "~/constants/api-constants";
-import type { IMovie } from "~/types/interfaces/IMovie";
-import type { IMovieResponse } from "~/types/interfaces/IMovieResponse";
+import type { IMovieandSeriesResponse } from "~/types/interfaces/IMovieandSeriesResponse";
 
 export function useMoviesAPI() {
-  // Fetch popular movies (Note: OMDb doesn't have a "popular" endpoint, so we search for common terms)
   const fetchMovies = (page = 1) => {
-    return $fetch<IMovieResponse>(`${API_BASE_URL}`, {
+    return $fetch<IMovieandSeriesResponse>(`${API_BASE_URL}`, {
       params: {
         apikey: API_KEY,
-        s: 'movie', // Search term, you can change this
+        s: 'movie',
         type: 'movie',
         page
       }
     });
   };
-  
-  // Fetch movie by ID
-  const fetchMovieDetails = (id: string) => {
-    return $fetch<IMovie>(`${API_BASE_URL}`, {
-      params: {
-        apikey: API_KEY,
-        i: id,
-        plot: 'full'
-      }
-    });
-  };
-  
+
   // Search movies
   const searchMovies = (query: string, page = 1) => {
-    return $fetch<IMovieResponse>(`${API_BASE_URL}`, {
+    return $fetch<IMovieandSeriesResponse>(`${API_BASE_URL}`, {
       params: {
         apikey: API_KEY,
         s: query,
@@ -40,7 +27,6 @@ export function useMoviesAPI() {
   
   return {
     fetchMovies,
-    fetchMovieDetails,
     searchMovies
   };
 }
