+++ composables/series-service.ts
@@ -1,11 +1,11 @@
 import { API_BASE_URL, API_KEY } from "~/constants/api-constants";
-import type { ISeries } from '~/types/interfaces/ISeries';
-import type { ISeriesResponse } from '~/types/interfaces/ISeriesResponse';
+import type { IMovieandSeries } from "~/types/interfaces/IMovieandSeries";
+import type { IMovieandSeriesResponse } from "~/types/interfaces/IMovieandSeriesResponse";
 
 export function useSeriesAPI() {
   // Fetch popular series
   const fetchSeries = (page = 1) => {
-    return $fetch<ISeriesResponse>(`${API_BASE_URL}`, {
+    return $fetch<IMovieandSeriesResponse>(`${API_BASE_URL}`, {
       params: {
         apikey: API_KEY,
         s: 'series', // Search term, you can change this
@@ -17,7 +17,7 @@ export function useSeriesAPI() {
   
   // Fetch series details
   const fetchSeriesDetails = (id: string) => {
-    return $fetch<ISeries>(`${API_BASE_URL}`, {
+    return $fetch<IMovieandSeries>(`${API_BASE_URL}`, {
       params: {
         apikey: API_KEY,
         i: id,
@@ -28,7 +28,7 @@ export function useSeriesAPI() {
   
   // Search series
   const searchSeries = (query: string, page = 1) => {
-    return $fetch<ISeriesResponse>(`${API_BASE_URL}`, {
+    return $fetch<IMovieandSeriesResponse>(`${API_BASE_URL}`, {
       params: {
         apikey: API_KEY,
         s: query,
