--- types/interfaces/IMovie.ts
@@ -1,14 +1,0 @@
-export interface IMovie {
-  Title: string;
-  Year: string;
-  imdbID: string;
-  Type: string;
-  Poster: string;
-  Plot?: string;
-  Released?: string;
-  Runtime?: string;
-  Genre?: string;
-  Director?: string;
-  Actors?: string;
-  imdbRating?: string;
-}
