--- types/interfaces/IMovieResponse.ts
@@ -1,9 +1,0 @@
-import type { IMovie } from "./IMovie";
-
-export interface IMovieResponse {
-    page: number;
-    results: IMovie[];
-    total_pages: number;
-    total_results: number;
-  }
-
