+++ types/interfaces/IMovieandSeriesResponse.ts
@@ -1,0 +1,9 @@
+import type { IMovieandSeries } from "./IMovieandSeries";
+
+export interface IMovieandSeriesResponse {
+    page: number;
+    results: IMovieandSeries[];
+    total_pages: number;
+    total_results: number;
+  }
+
